refactor: move paymaster code to a dedicated paymaster pkg
