+++ paymaster/types_paymaster.go
@@ -1,4 +1,4 @@
-package rpc
+package paymaster
 
 import (
 	"context"
